Move schema and schema-field code to type.ts
