+++ src/runtime/entity.ts
@@ -9,11 +9,10 @@
  */
 
 import {assert} from '../platform/assert-web.js';
-import {Type, EntityType, Schema, FieldType} from '../types/lib-types.js';
+import {EntityType, Schema, FieldType} from '../types/lib-types.js';
 import {Id, IdGenerator} from './id.js';
 import {Dictionary, Consumer} from '../utils/lib-utils.js';
 import {SYMBOL_INTERNALS} from './symbols.js';
-import {Refinement} from '../types/lib-types.js';
 import {Flags} from './flags.js';
 import {Ttl} from './capabilities.js';
 import {Storable, StorableInternals, getStorableInternals} from './storable.js';
