+++ src/types/internal/schema-from-literal.ts
@@ -8,10 +8,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {Schema} from './schema.js';
-import {Type, EntityType} from './type.js';
+import {Schema, Type, EntityType, FieldType} from './type.js';
 import {Refinement} from './refiner.js';
-import {FieldType} from './schema-field.js';
 import {SchemaFieldKind} from '../../runtime/manifest-ast-types/manifest-ast-nodes.js';
 
 function fromLiteral(data = {fields: {}, names: [], description: {}, refinement: null}) {
