+++ src/types/lib-types.ts
@@ -11,11 +11,8 @@
 export {Type, TypeLiteral, CountType, EntityType, SingletonType, CollectionType, BigCollectionType,
         ReferenceType, TupleType, MuxType, TypeVariable, TypeVariableInfo, InterfaceType, HandleType,
         HandleConnection, Slot, SlotType, InterfaceInfo, HandleConnectionLiteral, SlotLiteral,
-        TypeVarReference, InterfaceInfoLiteral, MatchResult} from './internal/type.js';
-
-export {Schema} from './internal/schema.js';
-export {FieldType, PrimitiveField, ReferenceField, InlineField, CollectionField, OrderedListField,
-        } from './internal/schema-field.js';
+        TypeVarReference, InterfaceInfoLiteral, MatchResult, Schema, FieldType, PrimitiveField,
+        ReferenceField, InlineField, CollectionField, OrderedListField } from './internal/type.js';
 
 export {Refinement, RefinementExpressionLiteral, RefinementExpressionVisitor, BinaryExpression,
         UnaryExpression, FieldNamePrimitive, QueryArgumentPrimitive, BuiltIn, NumberPrimitive,
